requirements.txt
torch
torchaudio
transformers
datasets
librosa
scikit-learn
numpy
pandas
plotly
matplotlib
seaborn
hydra-core
omegaconf